+++ src/types/gateway/ready.ts
@@ -1,5 +1,5 @@
 import { UnavailableGuild } from "../guilds/unavailable_guild.ts";
-import { Application } from "../oauth2/application.ts";
+import { Application } from "../applications/application.ts";
 import { User } from "../users/user.ts";
 import { SnakeCasedPropertiesDeep } from "../util.ts";
