+++ src/types/invites/invite.ts
@@ -1,6 +1,6 @@
 import { Channel } from "../channels/channel.ts";
 import { Guild } from "../guilds/guild.ts";
-import { Application } from "../oauth2/application.ts";
+import { Application } from "../applications/application.ts";
 import { User } from "../users/user.ts";
 import { DiscordTargetTypes } from "./target_types.ts";
